requirements.txt
requests
django
djangorestframework
python-dotenv